Extract some SymbolIsBannedAnalyzer functionality to base class
